add scripts to create kernel for Nanopi-M4
we will boot our kernel by upstream u-boot build for Nanopi-M4 2GB board.

* tools/mkimage: this tool is copied from u-boot and will create
the kernel image in a format which will be loaded by uboot.

* extlinux.conf: this is the boot configuration file needed by uboot to
find the kernel image and dtb on SD card.

* rk3399.dtb: dtb file is required by uboot the linux type of os so
added this dtb but we will not use the dtb in our project.

* make_kernel_image.sh: this script will create the final image which
can be flashed to SD card to run our kernel.

Signed-off-by: Deepak Das <email.deepakdas@gmail.com>